## Summary

Fixes the failing test "should return warning status for slow database response" in AdminService to achieve 100% test pass rate (88/88 tests passing).

### Issue Fixed

**Failing Test**: `AdminService › getSystemHealth › should return warning status for slow database response`
- **Expected**: Database status `"slow"`, System status `"warning"`
- **Received**: Database status `"connected"`, System status `"healthy"`
- **Root Cause**: `Date.now()` mock was configured for 3 calls but service only makes 2 calls

### Technical Solution

**Problem Analysis**:
The `getSystemHealth()` method in AdminService calls `Date.now()` exactly twice:
1. `const testStart = Date.now();` (line 139)
2. `responseTime = Date.now() - testStart;` (line 141)

The test was mocking 3 calls but only 2 were being made.

**Fix Applied**:
Updated `src/admin/admin.service.spec.ts` lines 230-235:

```typescript
// Before (incorrect - 3 call sequence)
if (callCount === 1) return 1000; // startTime
if (callCount === 2) return 1000; // testStart  
if (callCount === 3) return 2500; // testEnd

// After (correct - 2 call sequence)
if (callCount === 1) return 1000; // testStart
if (callCount === 2) return 2500; // testEnd (1500ms later, > 1000ms threshold)
```

### Test Results

**Before Fix**:
- ❌ Tests: 1 failed, 87 passed, 88 total
- AdminService test failing due to timing mock issue

**After Fix**:
- ✅ Tests: 88 passed, 88 total (100% pass rate achieved)
- All CI pipeline tests passing
- Deterministic and reliable test execution

### CI Validation

**All GitHub Actions Passing** ✅
- **Security Audit**: ✅ Passed (33s)
- **Backend Tests**: ✅ Passed (2m19s) - 88/88 tests including fixed admin test
- **Frontend Tests**: ✅ Passed (2m20s)
- **Performance Tests**: ✅ Passed (40s)
- **Smoke Tests**: ✅ Passed (11s)

### Quality Assurance

- **No service logic changed**: Only fixed test mock to match actual service behavior
- **Deterministic test**: Mock now accurately simulates >1000ms response time
- **Robust timing logic**: Properly validates slow database response detection
- **100% pass rate**: All tests now pass consistently

### Impact

This fix ensures:
- Complete test coverage validation
- Reliable CI/CD pipeline execution  
- Proper testing of system health monitoring features
- Foundation for future development with confidence in test suite

🤖 Generated with [Claude Code](https://claude.ai/code)